Проведён рефакторинг: Изменено название проекта, изменены названия тестов, изменено название главного класса.
