--- NeoflexBankProject/src/main/java/org/example/neoflexbankproject/NeoflexBankProjectApplication.java
@@ -1,13 +1,0 @@
-package org.example.neoflexbankproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NeoflexBankProjectApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NeoflexBankProjectApplication.class, args);
-    }
-
-}
