--- NeoflexBankProject/src/main/java/org/example/neoflexbankproject/service/CalculatorService.java
@@ -1,11 +1,0 @@
-package org.example.neoflexbankproject.service;
-
-import lombok.NonNull;
-import org.example.neoflexbankproject.dto.LoanOfferDto;
-import org.example.neoflexbankproject.dto.LoanStatementRequestDto;
-
-import java.util.List;
-
-public interface CalculatorService {
-    List<LoanOfferDto> salaryCalculation(@NonNull LoanStatementRequestDto requestDto);
-}
