--- NeoflexBankProject/src/main/java/org/example/neoflexbankproject/service/CreditCalculationService.java
@@ -1,8 +1,0 @@
-package org.example.neoflexbankproject.service;
-
-import org.example.neoflexbankproject.dto.CreditDto;
-import org.example.neoflexbankproject.dto.ScoringDataDto;
-
-public interface CreditCalculationService {
-    CreditDto calculateCredit(ScoringDataDto scoringDataDto);
-}
