+++ bankcalculator/src/main/java/org/example/bankcalculator/BankCalculatorApplication.java
@@ -1,0 +1,13 @@
+package org.example.bankcalculator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BankCalculatorApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BankCalculatorApplication.class, args);
+    }
+
+}
