+++ bankcalculator/src/main/java/org/example/bankcalculator/dto/EmailMessage.java
@@ -1,0 +1,13 @@
+package org.example.bankcalculator.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class EmailMessage {
+    private String address;
+    private Enum theme;               // Замените на конкретный enum, например EmailTheme
+    private Long statementId;
+    private String text;
+}
