+++ bankcalculator/src/main/java/org/example/bankcalculator/dto/LoanOfferDto.java
@@ -1,0 +1,20 @@
+package org.example.bankcalculator.dto;
+
+import lombok.Builder;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+@Builder
+public record LoanOfferDto(
+        UUID statementId,
+        BigDecimal requestedAmount,
+        BigDecimal totalAmount,
+        Integer term,
+        BigDecimal monthlyPayment,
+        BigDecimal rate,
+        Boolean isInsuranceEnabled,
+        Boolean isSalaryClient
+) {
+}
+
