+++ bankcalculator/src/main/java/org/example/bankcalculator/dto/StatementStatusHistoryDto.java
@@ -1,0 +1,14 @@
+package org.example.bankcalculator.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+public class StatementStatusHistoryDto {
+    private Enum status;              // Замените на конкретный enum, например StatementStatus
+    private LocalDateTime time;
+    private Enum changeType;          // Замените на конкретный enum, например ChangeType
+}
