+++ bankcalculator/src/main/java/org/example/bankcalculator/service/CalculatorService.java
@@ -1,0 +1,11 @@
+package org.example.bankcalculator.service;
+
+import lombok.NonNull;
+import org.example.bankcalculator.dto.LoanOfferDto;
+import org.example.bankcalculator.dto.LoanStatementRequestDto;
+
+import java.util.List;
+
+public interface CalculatorService {
+    List<LoanOfferDto> salaryCalculation(@NonNull LoanStatementRequestDto requestDto);
+}
