+++ bankcalculator/src/main/java/org/example/bankcalculator/service/CreditCalculationService.java
@@ -1,0 +1,8 @@
+package org.example.bankcalculator.service;
+
+import org.example.bankcalculator.dto.CreditDto;
+import org.example.bankcalculator.dto.ScoringDataDto;
+
+public interface CreditCalculationService {
+    CreditDto calculateCredit(ScoringDataDto scoringDataDto);
+}
