+++ bankcalculator/src/test/java/org/example/bankcalculator/BankCalculatorApplicationTests.java
@@ -1,0 +1,13 @@
+package org.example.bankcalculator;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+
+@AutoConfigureMockMvc
+@SpringBootTest
+public class BankCalculatorApplicationTests {
+    @Autowired
+    private MockMvc mockMvc;
+}
